requirements.txt
asn1crypto==0.24.0
astroid==2.2.5
atomicwrites==1.3.0
attrs==19.1.0
blinker==1.4
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
Click==7.0
codecov==2.0.15
coverage==4.5.3
cryptography==2.6.1
facebook-sdk==3.1.0
Flask==1.0.2
flask-mongoengine==0.9.5
Flask-WTF==0.14.2
gunicorn==19.9.0
idna==2.8
isort==4.3.17
itsdangerous==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.1.1
mccabe==0.6.1
mongoengine==0.17.0
mongomock==3.16.0
more-itertools==6.0.0
pluggy==0.9.0
py==1.8.0
pycparser==2.19
pylint==2.3.1
pymongo==3.7.2
pytest==4.3.1
pytest-cov==2.6.1
pytest-flask==0.14.0
python-http-client==3.1.0
requests==2.21.0
sendgrid==6.0.4
sentinels==1.0.0
six==1.12.0
typed-ast==1.3.1
urllib3==1.24.1
Werkzeug==0.15.1
wrapt==1.11.1
WTForms==2.2.1
